requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.1            # via django
django-environ==0.4.5     # via -r requirements.in
django-ipware==3.0.2      # via -r requirements.in
django==3.1.4             # via -r requirements.in
pytz==2020.4              # via django
sqlparse==0.4.1           # via django
